+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/wascore/perfcounter/jvm/DeadLockDetectorPerformanceCounter.java
@@ -32,12 +32,12 @@ import com.azure.monitor.opentelemetry.exporter.implementation.models.TelemetryI
 import com.microsoft.applicationinsights.agent.internal.wascore.FormattedTime;
 import com.microsoft.applicationinsights.agent.internal.wascore.TelemetryClient;
 import com.microsoft.applicationinsights.agent.internal.wascore.perfcounter.PerformanceCounter;
-import com.microsoft.applicationinsights.agent.internal.wascore.util.LocalStringsUtils;
 import java.lang.management.ManagementFactory;
 import java.lang.management.MonitorInfo;
 import java.lang.management.ThreadInfo;
 import java.lang.management.ThreadMXBean;
 import java.util.ArrayList;
+import java.util.concurrent.ThreadLocalRandom;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -53,8 +53,6 @@ public final class DeadLockDetectorPerformanceCounter implements PerformanceCoun
   private static final Logger logger =
       LoggerFactory.getLogger(DeadLockDetectorPerformanceCounter.class);
 
-  public static final String NAME = "ThreadDeadLockDetector";
-
   private static final String INDENT = "    ";
   private static final String SEPERATOR = " | ";
   private static final String METRIC_NAME = "Suspected Deadlocked Threads";
@@ -64,11 +62,6 @@ public final class DeadLockDetectorPerformanceCounter implements PerformanceCoun
 
   public DeadLockDetectorPerformanceCounter() {
     threadBean = ManagementFactory.getThreadMXBean();
-  }
-
-  public boolean isSupported() {
-    ThreadMXBean threadBean = ManagementFactory.getThreadMXBean();
-    return threadBean.isSynchronizerUsageSupported();
   }
 
   @Override
@@ -103,7 +96,7 @@ public final class DeadLockDetectorPerformanceCounter implements PerformanceCoun
       }
 
       if (!blockedThreads.isEmpty()) {
-        String uuid = LocalStringsUtils.generateRandomIntegerId();
+        String uuid = generateRandomIntegerId();
 
         data.getMetrics().get(0).setValue(blockedThreads.size());
         telemetry.getTags().put(ContextTagKeys.AI_OPERATION_ID.toString(), uuid);
@@ -179,5 +172,11 @@ public final class DeadLockDetectorPerformanceCounter implements PerformanceCoun
           .append(" Id=")
           .append(ti.getLockOwnerId());
     }
+  }
+
+  private static String generateRandomIntegerId() {
+    // avoid using Math.abs(rand.nextLong()) because Math.abs(Long.MIN_VALUE) is negative
+    long rand = ThreadLocalRandom.current().nextLong(Long.MAX_VALUE);
+    return String.valueOf(rand);
   }
 }